Modify Scripts and Added Scripts
Added: SpawnManager, Assault, EnemyManager
Modify: Health, EnemyController(used to be Enemy_Function), PlayerAI, Caravan, InteractWithCaravan

Features:
- Enemy can now attack caravan and breaks parts.
- Added attact state and conditions for PlayerAI.

+++ Catavaneers/Assets/Scripts/Caravan/Part.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Part : MonoBehaviour
 {
-    [SerializeField] int health_value = 2;
-    public int healthValue { get { return health_value; } }
+    [SerializeField] float health_value = 2f;
+    public float healthValue { get { return health_value; } }
 
     // Property field
     /*
